chore: add jsonargparse and dataclasses to manage configs
