+++ requirements.txt
@@ -1,4 +1,7 @@
-tensorflow == 2.17.0
+tensorflow==2.17.0
+pandas==2.2.2
+opencv-python==4.10.0.84
 PyYAML~=6.0.2rc1
 pytest~=6.2.5
-jsonargparse == 4.32.0
+jsonargparse==4.32.0
+scikit-learn==1.3.2
