+++ tests/conftest.py
@@ -1,0 +1,3 @@
+pytest_plugins = [
+    "tests.model_fixtures"
+]
